requirements.txt
python-dotenv==0.18.0
slack-sdk==3.7.0
